README.md
Hi! 

So this was a very quick iOS app I made on the fly using some files from a YouTuber. 

All the images that were used were his. I simply coded it so that it would work. 

This was a very prompt project as I was in the airport and didn't have enough WiFi, and I was extremely bored. Hah! 

Plan on making more projects like this soon.